Explore resolver invalidation
