+++ packages/liveblocks-core/src/lib/batch.ts
@@ -16,6 +16,7 @@ export type BatchCallback<O, I> = (
 export type BatchStore<O, I> = Observable<void> & {
   get: (input: I) => Promise<void>;
   getState: (input: I) => AsyncResult<O> | undefined;
+  invalidate: (input: I) => void;
 };
 
 interface Options {
@@ -170,6 +171,16 @@ export function createBatchStore<O, I>(batch: Batch<O, I>): BatchStore<O, I> {
     eventSource.notify();
   }
 
+  function invalidate(input: I): void {
+    const cacheKey = getCacheKey(input);
+
+    // Clear the cache for this call.
+    cache.delete(cacheKey);
+
+    // Notify subscribers.
+    eventSource.notify();
+  }
+
   async function get(input: I): Promise<void> {
     const cacheKey = getCacheKey(input);
 
@@ -220,5 +231,6 @@ export function createBatchStore<O, I>(batch: Batch<O, I>): BatchStore<O, I> {
     ...eventSource.observable,
     get,
     getState,
+    invalidate,
   };
 }
