+++ packages/liveblocks-react-lexical/src/mentions/mention-plugin.tsx
@@ -10,8 +10,7 @@ import {
 } from "@floating-ui/react-dom";
 import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
 import { kInternal } from "@liveblocks/core";
-import { useRoom } from "@liveblocks/react";
-import { useMentionSuggestions } from "@liveblocks/react-ui";
+import { useMentionSuggestions, useRoom } from "@liveblocks/react";
 import type { EditorState, NodeKey, NodeMutation, TextNode } from "lexical";
 import {
   $createRangeSelection,
