+++ packages/liveblocks-react-ui/src/index.ts
@@ -41,5 +41,4 @@ export type {
 export { useOverrides } from "./overrides";
 export type { ComposerSubmitComment } from "./primitives";
 export { Timestamp } from "./primitives/Timestamp";
-export { useMentionSuggestions } from "./shared";
 export type { CommentAttachmentArgs } from "./types";
